services/rag/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
structlog==24.1.0
httpx==0.27.0
anthropic==0.25.1
openai==1.14.3
pydantic==2.6.4
python-dotenv==1.0.1
sqlalchemy==2.0.29
asyncpg
pgvector
